requirements.txt
# 医药连锁名称关联 Agent - 依赖包
# Python 3.14 兼容版本

# 核心依赖
python-dotenv>=1.0.0
pyyaml>=6.0.1

# 数据处理（使用最新版本以支持 Python 3.14）
pandas>=2.2.0
openpyxl>=3.1.2
xlsxwriter>=3.1.9

# 数据库
sqlalchemy>=2.0.25

# LLM 客户端
anthropic>=0.18.1

# 网络请求
requests>=2.31.0

# 工具库
python-dateutil>=2.8.2
tzlocal>=5.2

# 命令行界面
rich>=13.7.0
prompt-toolkit>=3.0.43

# 日志
colorlog>=6.8.0

# 测试（可选）
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# 代码质量（可选）
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
pylint>=3.0.3
